issue #725. 1st working version.
