+++ src/psyclone/psyir/transformations/intrinsics/__init__.py
@@ -1,7 +1,7 @@
 # -----------------------------------------------------------------------------
 # BSD 3-Clause License
 #
-# Copyright (c) 2019, Science and Technology Facilities Council.
+# Copyright (c) 2020, Science and Technology Facilities Council
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -31,37 +31,11 @@
 # ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 # -----------------------------------------------------------------------------
-# Authors J. Henrichs, Bureau of Meteorology
+# Author R. W. Ford, STFC Daresbury Lab
 
-'''Transformation module, containing all generic (API independent)
-transformations and base classes.
+'''PSyclone Internal Representation intrinsics transformation
+module. Contains all transformations that act on PSyIR intrinsic
+(operator) functions.
+
 '''
-
-from psyclone.psyir.transformations.extract_trans import ExtractTrans
-from psyclone.psyir.transformations.profile_trans import ProfileTrans
-from psyclone.psyir.transformations.psy_data_trans import PSyDataTrans
-from psyclone.psyir.transformations.region_trans import RegionTrans
-from psyclone.psyir.transformations.intrinsics.abs2code_trans import \
-    Abs2CodeTrans
-from psyclone.psyir.transformations.intrinsics.matmul2code_trans import \
-    Matmul2CodeTrans
-from psyclone.psyir.transformations.intrinsics.min2code_trans import \
-    Min2CodeTrans
-from psyclone.psyir.transformations.intrinsics.sign2code_trans import \
-    Sign2CodeTrans
-from psyclone.psyir.transformations.transformation_error \
-    import TransformationError
-
-# The entities in the __all__ list are made available to import directly from
-# this package e.g.:
-# from psyclone.psyir.transformations import ExtractTrans
-
-__all__ = ['ExtractTrans',
-           'ProfileTrans',
-           'PSyDataTrans',
-           'RegionTrans',
-           'Abs2CodeTrans',
-           'Matmul2CodeTrans',
-           'Min2CodeTrans',
-           'Sign2CodeTrans',
-           'TransformationError']
+__all__ = []
