+++ src/psyclone/psyir/transformations/intrinsics/operator2code_trans.py
@@ -49,16 +49,16 @@ from psyclone.psyir.transformations.transformation_error import \
 
 
 @six.add_metaclass(abc.ABCMeta)
-class NemoOperatorTrans(Transformation):
-    '''Provides NEMO-API-specific support for transformations from PSyIR
-    intrinsic Operator nodes to equivalent PSyIR code in a PSyIR
-    tree. Such transformations can be useful when the intrinsic is not
-    supported by a particular backend or if it is more efficient to
-    have explicit code.
+class Operator2CodeTrans(Transformation):
+    '''Provides support for transformations from PSyIR intrinsic Operator
+    nodes to equivalent PSyIR code in a PSyIR tree. Such
+    transformations can be useful when the intrinsic is not supported
+    by a particular backend or if it is more efficient to have
+    explicit code.
 
     '''
     def __init__(self):
-        super(NemoOperatorTrans, self).__init__()
+        super(Operator2CodeTrans, self).__init__()
         self._operator_name = None
         self._classes = None
         self._operators = None
@@ -74,7 +74,7 @@ class NemoOperatorTrans(Transformation):
         :rtype:str
 
         '''
-        return "Nemo{0}Trans".format(self._operator_name.title())
+        return "{0}2CodeTrans".format(self._operator_name.title())
 
     def validate(self, node, options=None):
         '''Perform various checks to ensure that it is valid to apply
@@ -125,6 +125,3 @@ class NemoOperatorTrans(Transformation):
     def apply(self, node, options=None):
         '''Abstract method, see psyclone.psyGen.Transformations apply() for
         more details.'''
-        schedule = None
-        memento = None
-        return schedule, memento
