+++ src/psyclone/psyir/transformations/intrinsics/sign2code_trans.py
@@ -31,146 +31,151 @@
 # ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 # -----------------------------------------------------------------------------
-# Author: R. W. Ford, STFC Daresbury Lab
+# Author: R. W. Ford, STFC Daresbury Laboratory
 # Modified: A. R. Porter, STFC Daresbury Laboratory
 
-'''Module providing a NEMO-api-specific transformation from a PSyIR
-ABS operator to PSyIR code. This could be useful if the ABS operator
-is not supported by the back-end or if the performance in the inline
-code is better than the intrinsic.
-
-This implementation is no longer NEMO-specific and should be modified
-(and renamed) to work for all APIs (#725).
+'''Module providing a transformation from a PSyIR SIGN operator to
+PSyIR code. This could be useful if the SIGN operator is not supported
+by the back-end or if the performance of the inline code is better
+than the intrinsic.
 
 '''
 from __future__ import absolute_import
-from psyclone.undoredo import Memento
-from psyclone.psyir.transformations.nemo_operator_trans import \
-    NemoOperatorTrans
+from psyclone.psyir.transformations.intrinsics.operator2code_trans import \
+    Operator2CodeTrans
+from psyclone.psyir.transformations import Abs2CodeTrans
 from psyclone.psyir.nodes import UnaryOperation, BinaryOperation, Assignment, \
     Reference, Literal, IfBlock
 from psyclone.psyir.symbols import DataSymbol, REAL_TYPE
 
 
-class NemoAbsTrans(NemoOperatorTrans):
-    '''Provides a NEMO-api-specific transformation from a PSyIR ABS
-    Operator node to equivalent code in a PSyIR tree. Validity checks
-    are also performed.
+class Sign2CodeTrans(Operator2CodeTrans):
+    '''Provides a transformation from a PSyIR SIGN Operator node to
+    equivalent code in a PSyIR tree. Validity checks are also
+    performed.
 
     The transformation replaces
 
     .. code-block:: python
 
-        R = ABS(X)
+        R = SIGN(A, B)
 
     with the following logic:
 
     .. code-block:: python
 
-        IF X < 0.0:
-            R = X*-1.0
-        ELSE:
-            R = X
+        R = ABS(A)
+        if B < 0.0:
+            R = R*-1.0
+
+    i.e. the value of ``A`` with the sign of ``B``
 
     '''
     def __init__(self):
-        super(NemoAbsTrans, self).__init__()
-        self._operator_name = "ABS"
-        self._classes = (UnaryOperation,)
-        self._operators = (UnaryOperation.Operator.ABS,)
+        super(Sign2CodeTrans, self).__init__()
+        self._operator_name = "SIGN"
+        self._classes = (BinaryOperation,)
+        self._operators = (BinaryOperation.Operator.SIGN,)
 
     def apply(self, node, options=None):
-        '''Apply the ABS intrinsic conversion transformation to the specified
-        node. This node must be an ABS UnaryOperation. The ABS
-        UnaryOperation is converted to equivalent inline code. This is
-        implemented as a PSyIR transform from:
+        '''Apply the SIGN intrinsic conversion transformation to the specified
+        node. This node must be a SIGN BinaryOperation. The SIGN
+        BinaryOperation is converted to equivalent inline code. This
+        is implemented as a PSyIR transform from:
 
         .. code-block:: python
 
-            R = ... ABS(X) ...
+            R = ... SIGN(A, B) ...
 
         to:
 
         .. code-block:: python
 
-            tmp_abs = X
+            tmp_abs = A
             if tmp_abs < 0.0:
                 res_abs = tmp_abs*-1.0
             else:
                 res_abs = tmp_abs
-            R = ... res_abs ...
+            res_sign = res_abs
+            tmp_sign = B
+            if tmp_sign < 0.0:
+                res_sign = res_sign*-1.0
+            R = ... res_sign ...
 
-        where ``X`` could be an arbitrarily complex PSyIR expression
-        and ``...`` could be arbitrary PSyIR code.
+        where ``A`` and ``B`` could be arbitrarily complex PSyIR
+        expressions, ``...`` could be arbitrary PSyIR code and where
+        ``ABS`` has been replaced with inline code by the NemoAbsTrans
+        transformation.
 
         This transformation requires the operation node to be a
         descendent of an assignment and will raise an exception if
         this is not the case.
 
-        :param node: an ABS UnaryOperation node.
-        :type node: :py:class:`psyclone.psyGen.UnaryOperation`
+        :param node: a SIGN BinaryOperation node.
+        :type node: :py:class:`psyclone.psyGen.BinaryOperation`
+        :param symbol_table: the symbol table.
+        :type symbol_table: :py:class:`psyclone.psyir.symbols.SymbolTable`
         :param options: a dictionary with options for transformations.
         :type options: dictionary of string:values or None
-
-        :returns: 2-tuple of new schedule and memento of transform.
-        :rtype: (:py:class:`psyclone.nemo.NemoInvokeSchedule`, \
-                 :py:class:`psyclone.undoredo.Memento`)
 
         '''
         self.validate(node)
 
         schedule = node.root
         symbol_table = schedule.symbol_table
-        memento = Memento(schedule, self, [node])
 
         oper_parent = node.parent
         assignment = node.ancestor(Assignment)
         # Create two temporary variables.  There is an assumption here
-        # that the ABS Operator returns a PSyIR real type. This might
+        # that the SIGN Operator returns a PSyIR real type. This might
         # not be what is wanted (e.g. the args might PSyIR integers),
         # or there may be errors (arguments are of different types)
         # but this can't be checked as we don't have the appropriate
         # methods to query nodes (see #658).
-        res_var = symbol_table.new_symbol_name("res_abs")
-        symbol_res_var = DataSymbol(res_var, REAL_TYPE)
-        symbol_table.add(symbol_res_var)
-        tmp_var = symbol_table.new_symbol_name("tmp_abs")
-        symbol_tmp_var = DataSymbol(tmp_var, REAL_TYPE)
-        symbol_table.add(symbol_tmp_var)
+        res_var = symbol_table.new_symbol_name("res_sign")
+        res_var_symbol = DataSymbol(res_var, REAL_TYPE)
+        symbol_table.add(res_var_symbol)
+        tmp_var = symbol_table.new_symbol_name("tmp_sign")
+        tmp_var_symbol = DataSymbol(tmp_var, REAL_TYPE)
+        symbol_table.add(tmp_var_symbol)
 
-        # Replace operation with a temporary (res_X).
-        oper_parent.children[node.position] = Reference(symbol_res_var,
+        # Replace operator with a temporary (res_var).
+        oper_parent.children[node.position] = Reference(res_var_symbol,
                                                         parent=oper_parent)
 
-        # tmp_var=X
-        lhs = Reference(symbol_tmp_var)
-        rhs = node.children[0]
+        # res_var=ABS(A)
+        lhs = Reference(res_var_symbol)
+        rhs = UnaryOperation.create(UnaryOperation.Operator.ABS,
+                                    node.children[0])
         new_assignment = Assignment.create(lhs, rhs)
         new_assignment.parent = assignment.parent
         assignment.parent.children.insert(assignment.position, new_assignment)
 
-        # if condition: tmp_var>0.0
-        lhs = Reference(symbol_tmp_var)
+        # Replace the ABS intrinsic with inline code.
+        abs_trans = Abs2CodeTrans()
+        abs_trans.apply(rhs, symbol_table)
+
+        # tmp_var=B
+        lhs = Reference(tmp_var_symbol)
+        new_assignment = Assignment.create(lhs, node.children[1])
+        new_assignment.parent = assignment.parent
+        assignment.parent.children.insert(assignment.position, new_assignment)
+
+        # if_condition: tmp_var<0.0
+        lhs = Reference(tmp_var_symbol)
         rhs = Literal("0.0", REAL_TYPE)
-        if_condition = BinaryOperation.create(BinaryOperation.Operator.GT,
+        if_condition = BinaryOperation.create(BinaryOperation.Operator.LT,
                                               lhs, rhs)
 
-        # then_body: res_var=tmp_var
-        lhs = Reference(symbol_res_var)
-        rhs = Reference(symbol_tmp_var)
+        # then_body: res_var=res_var*-1.0
+        lhs = Reference(res_var_symbol)
+        lhs_child = Reference(res_var_symbol)
+        rhs_child = Literal("-1.0", REAL_TYPE)
+        rhs = BinaryOperation.create(BinaryOperation.Operator.MUL,
+                                     lhs_child, rhs_child)
         then_body = [Assignment.create(lhs, rhs)]
 
-        # else_body: res_var=-1.0*tmp_var
-        lhs = Reference(symbol_res_var)
-        lhs_child = Reference(symbol_tmp_var)
-        rhs_child = Literal("-1.0", REAL_TYPE)
-        rhs = BinaryOperation.create(BinaryOperation.Operator.MUL, lhs_child,
-                                     rhs_child)
-        else_body = [Assignment.create(lhs, rhs)]
-
-        # if [if_condition] then [then_body] else [else_body]
-        if_stmt = IfBlock.create(if_condition, then_body, else_body)
+        # if [if_condition] then [then_body]
+        if_stmt = IfBlock.create(if_condition, then_body)
         if_stmt.parent = assignment.parent
         assignment.parent.children.insert(assignment.position, if_stmt)
-
-        return schedule, memento
