+++ src/psyclone/tests/psyir/transformations/intrinsics/sign2code_trans_test.py
@@ -33,11 +33,11 @@
 # -----------------------------------------------------------------------------
 # Author R. W. Ford, STFC Daresbury Lab
 
-'''Module containing tests for the nemo abs transformation.'''
+'''Module containing tests for the abs2code transformation.'''
 
 from __future__ import absolute_import
 import pytest
-from psyclone.psyir.transformations import NemoSignTrans, TransformationError
+from psyclone.psyir.transformations import Sign2CodeTrans, TransformationError
 from psyclone.psyir.symbols import SymbolTable, DataSymbol, \
     ArgumentInterface, REAL_TYPE
 from psyclone.psyir.nodes import Reference, BinaryOperation, Assignment, \
@@ -53,13 +53,13 @@ def test_initialise():
     class is created and that the str and name methods work as expected.
 
     '''
-    trans = NemoSignTrans()
+    trans = Sign2CodeTrans()
     assert trans._operator_name == "SIGN"
     assert trans._classes == (BinaryOperation,)
     assert trans._operators == (BinaryOperation.Operator.SIGN,)
     assert (str(trans) == "Convert the PSyIR SIGN intrinsic to equivalent "
             "PSyIR code.")
-    assert trans.name == "NemoSignTrans"
+    assert trans.name == "Sign2CodeTrans"
 
 
 def example_psyir(create_expression):
@@ -118,8 +118,8 @@ def test_correct(func, output, tmpdir):
         "  real :: psyir_tmp\n\n"
         "  psyir_tmp=SIGN({0}, arg_1)\n\n"
         "end subroutine sign_example\n".format(output)) in result
-    trans = NemoSignTrans()
-    _, _ = trans.apply(operation, operation.root.symbol_table)
+    trans = Sign2CodeTrans()
+    trans.apply(operation, operation.root.symbol_table)
     result = writer(operation.root)
     assert (
         "subroutine sign_example(arg,arg_1)\n"
@@ -174,8 +174,8 @@ def test_correct_expr(tmpdir):
         "  real :: psyir_tmp\n\n"
         "  psyir_tmp=1.0 + SIGN(arg * 3.14, arg_1) + 2.0\n\n"
         "end subroutine sign_example\n") in result
-    trans = NemoSignTrans()
-    _, _ = trans.apply(operation, operation.root.symbol_table)
+    trans = Sign2CodeTrans()
+    trans.apply(operation, operation.root.symbol_table)
     result = writer(operation.root)
     assert (
         "subroutine sign_example(arg,arg_1)\n"
@@ -231,9 +231,9 @@ def test_correct_2sign(tmpdir):
         "  real :: psyir_tmp\n\n"
         "  psyir_tmp=SIGN(1.0, 1.0) + SIGN(arg * 3.14, arg_1)\n\n"
         "end subroutine sign_example\n") in result
-    trans = NemoSignTrans()
-    _, _ = trans.apply(operation, operation.root.symbol_table)
-    _, _ = trans.apply(sign_op, operation.root.symbol_table)
+    trans = Sign2CodeTrans()
+    trans.apply(operation, operation.root.symbol_table)
+    trans.apply(sign_op, operation.root.symbol_table)
     result = writer(operation.root)
     assert (
         "subroutine sign_example(arg,arg_1)\n"
@@ -282,11 +282,11 @@ def test_invalid():
     called.'''
     Config.get().api = "dynamo0.3"
     operation = example_psyir(lambda arg: arg)
-    trans = NemoSignTrans()
+    trans = Sign2CodeTrans()
     with pytest.raises(TransformationError) as excinfo:
-        _, _ = trans.apply(operation, operation.root.symbol_table)
+        trans.apply(operation, operation.root.symbol_table)
     assert (
-        "Error in NemoSignTrans transformation. This transformation only "
+        "Error in Sign2CodeTrans transformation. This transformation only "
         "works for the nemo API, but found 'dynamo0.3'"
         in str(excinfo.value))
     # Remove the created config instance
